demo-target/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pyjwt==2.8.0
